@dwt 2025-05-11
Waiting on https://github.com/NixOS/nixpkgs/pull/384691