@felixsinger 2025-06-16
> Upstream uses that option on their build server, so we should too.

Same here.. Please don't use "because upstream does thing x too" as a reason. "For compilation speed", which you did in the commit title, is enough.

It's clear that there is a significant improvement on compilation speed. So I've compared the times with #405384 applied in addition.

Without Ninja

```
real	2m51,167s
user	0m5,613s
sys	0m2,144s
```

With Ninja

```
real	2m40,752s
user	0m4,875s
sys	0m2,023s
```