@dwt 2025-06-19
Unity builds in contrast to #405384 reduce build times much more significantly for me.

From ~73s to ~42s, almost halving the build time. This is without the ninja change, which seems to compound nicely (as your research seems to show).

```shell
# without patch
❯ nx hyperfine 'nix build .#mumble --rebuild' --warmup 1 --ignore-failure
Benchmark 1: nix build .#mumble --rebuild
  Time (mean ± σ):     73.287 s ±  2.462 s    [User: 0.679 s, System: 3.338 s]
  Range (min … max):   69.478 s … 77.686 s    10 runs
 
  Warning: Ignoring non-zero exit code.

# with patch
❯ nx hyperfine 'nix build .#mumble --rebuild' --warmup 1 --ignore-failure
Benchmark 1: nix build .#mumble --rebuild
  Time (mean ± σ):     42.713 s ±  1.435 s    [User: 0.089 s, System: 1.216 s]
  Range (min … max):   39.891 s … 44.800 s    10 runs
 
  Warning: Ignoring non-zero exit code.
```