@dwt 2025-07-12
Friendly ping @felixsinger. Do you like the speedup produced by this pull request? Do you plan to merge it?